+++ HISTORY.md
@@ -1,6 +1,14 @@
 Sanitize History
 ================================================================================
 
+Version 2.1.0 (git)
+-------------------
+
+* Added support for whitelisting arbitrary HTML5 `data-*` attributes. Use the
+  symbol `:data` instead of an attribute name in the `:attributes` config to
+  indicate that arbitrary data attributes should be allowed on an element.
+
+
 Version 2.0.6 (2013-07-10)
 --------------------------